backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
pydantic
python-dotenv
requests
python-jose[cryptography]
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2